components split seprate
