+++ apps/gotask_ui/src/app/(portal)/goals/projectid/[projectId]/components/projectGoal/GoalCard.tsx
@@ -1,0 +1,69 @@
+// components/goal/GoalCard.tsx
+
+import React from "react";
+import { Box, Typography, IconButton, Tooltip } from "@mui/material";
+import { Edit } from "@mui/icons-material";
+import StatusIndicator from "@/app/component/status/statusIndicator";
+import { getStatusColor } from "@/app/common/constants/task";
+
+export interface GoalCardProps {
+  goal: {
+    id: string;
+    goalTitle: string;
+    status: string;
+  };
+  onEdit: (goal: any) => void;
+}
+
+const GoalCard: React.FC<GoalCardProps> = ({ goal, onEdit }) => {
+  const color = getStatusColor(goal.status);
+
+  return (
+    <Box
+      sx={{
+        p: 2,
+        borderRadius: 2,
+        backgroundColor: `${color}12`,
+        border: `1px solid ${color}40`,
+        transition: "all 0.2s ease-in-out",
+        cursor: "default",
+        display: "flex",
+        flexDirection: "column",
+        height: "100%",
+        justifyContent: "space-between",
+        "&:hover": {
+          boxShadow: `0 0 6px ${color}66`
+        }
+      }}
+    >
+      {/* Goal Title and Status */}
+      <Box>
+        <Typography variant="subtitle1" fontWeight={500} textTransform="capitalize" sx={{ mb: 1 }}>
+          {goal.goalTitle}
+        </Typography>
+
+        <StatusIndicator status={goal.status} getColor={getStatusColor} />
+
+        {goal.status?.toLowerCase() === "completed" && (
+          <Box display="flex" alignItems="center" mt={1}>
+            <Typography variant="body2" sx={{ mr: 1 }}>
+              Completion:
+            </Typography>
+            <input type="checkbox" disabled checked />
+          </Box>
+        )}
+      </Box>
+
+      {/* Edit Button */}
+      <Box display="flex" justifyContent="flex-end" mt={2}>
+        <Tooltip title="Edit Goal">
+          <IconButton onClick={() => onEdit(goal)} color="primary">
+            <Edit />
+          </IconButton>
+        </Tooltip>
+      </Box>
+    </Box>
+  );
+};
+
+export default GoalCard;
